fixed integration making an internal download endpoint
