+++ daas/daas_app/api/__init__.py
@@ -2,4 +2,5 @@ from .get_sample_from_hash import GetSampleFromHashAPIView
 from .get_token import get_token_view
 from .reprocess import ReprocessAPIView
 from .upload import UploadAPIView
+from .download_sample import download_sample_view
 from . import internals
