+++ daas/daas_app/api/internals/__init__.py
@@ -1,3 +1,4 @@
 from .set_result import SetResultApiView
 from .chart_data import (SamplesPerSizeData, SamplesPerElapsedTimeData, SamplesPerTypeData, SamplesPerProcessDateData,
                          SamplesPerStatusForFileTypeData, SamplesPerUploadDateData)
+from .download_sample import download_sample_view
