+++ daas/daas_app/views/other_views.py
@@ -81,17 +81,6 @@ def no_filter_found_view(request):
 
 
 @login_required
-@permission_required('download_sample_permission')
-def download_sample_view(request, sample_id):
-    logging.info('downloading sample: id=%s' % sample_id)
-    sample = Sample.objects.get(id=sample_id)
-    # With the following 'if' nobody will be allowed to download samples if the config say so,
-    # even if they manually craft a download url.
-    file_content = sample.data if ALLOW_SAMPLE_DOWNLOAD else b''
-    return download(file_content, sample.name, "application/octet-stream")
-
-
-@login_required
 @permission_required('download_source_code_permission')
 def download_source_code_view(request, sample_id):
     logging.info('downloading source code: sample_id=%s' % sample_id)
